JITL-321

+++ src/main/java/com/sos/scheduler/notification/db/DBItemNotificationSchedulerHistoryOrderStep.java
@@ -36,7 +36,7 @@ public class DBItemNotificationSchedulerHistoryOrderStep implements Serializable
     private String taskCause;
     private Date taskStartTime;
     private Date taskEndTime;
-    private Long taskExitCode;
+    private Integer taskExitCode;
     
     public DBItemNotificationSchedulerHistoryOrderStep() {
     }
@@ -228,12 +228,12 @@ public class DBItemNotificationSchedulerHistoryOrderStep implements Serializable
     }
 
     @Column(name = "`TASK_EXIT_CODE`", nullable = true)
-    public Long getTaskExitCode() {
+    public Integer getTaskExitCode() {
         return taskExitCode;
     }
 
     @Column(name = "`TASK_EXIT_CODE`", nullable = true)
-    public void setTaskExitCode(Long val) {
+    public void setTaskExitCode(Integer val) {
         this.taskExitCode = val;
     }
